docker/aiagentui/requirements.txt
flask
openai
requests
docker
werkzeug
python-dotenv  
